Move moment calculator from SubpocketsFeature class to kissim.utils
