+++ kissim/tests/encoding/test_features_subpockets.py
@@ -250,25 +250,3 @@ class TestsSubpocketsFeature:
             assert np.isnan(distance_calculated)
         else:
             assert pytest.approx(distance_calculated, abs=1e-6) == distance
-
-    @pytest.mark.parametrize(
-        "values, moments",
-        [
-            ([0], [0, 0, 0]),
-            ([0, 0], [0, 0, 0]),
-            ([1, 0], [0.5, 0.5, 0]),
-            ([3, 0, 0], [1, 1.4142135, 1.2599210]),
-            ([], [np.nan, np.nan, np.nan]),
-            ([np.nan, np.nan], [np.nan, np.nan, np.nan]),
-        ],
-    )
-    def test_calculate_first_second_third_moment(self, values, moments):
-        """
-        Test static method that calculates the first three moments of a distribution.
-        """
-        feature = SubpocketsFeature()
-        moments_calculated = feature.calculate_first_second_third_moments(values)
-        if len(values) > 0 and not all(np.isnan(values)):
-            assert pytest.approx(moments_calculated, abs=1e-6) == moments
-        else:
-            assert all(moments_calculated) == all(moments)
