+++ kissim/utils.py
@@ -9,8 +9,11 @@ import os
 import shutil
 import tempfile
 import contextlib
-
 from multiprocessing import cpu_count
+
+import numpy as np
+from scipy.special import cbrt
+from scipy.stats.stats import moment
 
 
 logger = logging.getLogger(__name__)
@@ -63,3 +66,29 @@ def set_n_cores(n_cores):
             )
     logger.info(f"Number of cores used: {n_cores}.")
     return n_cores
+
+
+def calculate_first_second_third_moments(
+    values,
+):
+    """
+    Get first, second, and third moment (mean, standard deviation, and skewness)
+    for a distribution of values.
+    Note: Moments are based only on non-NaN values.
+
+    Parameters
+    ----------
+    values : list or numpy.array or pd.Series of (float or int)
+        List of values.
+    """
+
+    values = np.array(values)
+
+    if len(values) > 0 and not all(np.isnan(values)):
+        moment1 = np.nanmean(values)
+        # Second and third moment: delta degrees of freedom = 0 (divisor N)
+        moment2 = np.nanstd(values, ddof=0)
+        moment3 = cbrt(moment(values, moment=3, nan_policy="omit"))
+        return moment1, moment2, moment3
+    else:
+        return np.nan, np.nan, np.nan
